% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{download_microdata}
\alias{download_microdata}
\title{Download school census microdata}
\usage{
download_microdata(year = 2019, method = "wget")
}
\arguments{
\item{year}{numeric, at the moment works with 2018 and 2019 data}

\item{method}{character, default is "wget". Other options: "internal", "libcurl", "curl"}
}
\value{
.zip file with all microdata, which must be unzipped
}
\description{
Download school census microdata
}
\details{
This function works only with 2019 and 2018 data.

Other years can be downloaded in: http://inep.gov.br/microdados

This package works fine with 2019 data, and at least partially with 2015 data. Other years were not tested till now.
}
\examples{
\donttest{
download_microdata()
download_microdata(year=2018)}

# for a more complete view, see ?insert_labels
}
