% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{import_csv2rda}
\alias{import_csv2rda}
\title{Import csv microdata to R}
\usage{
import_csv2rda(file_path_origin, file_path_destiny)
}
\arguments{
\item{file_path_origin}{character, path to downloaded .csv file}

\item{file_path_destiny}{character, path to .rda file to be created}
}
\value{
data.frame saved in .rda file
}
\description{
Import csv microdata to R, optimizing for big data (student and teacher levels)
}
\examples{
\dontrun{
file.downloaded='~/YOUR_PATH/downloaded_data/ESCOLA.CSV'
file.imported='~/YOUR_PATH/temp_data.rda'

import_csv2rda(file_path_origin=file.downloaded,
              file_path_destiny=file.imported)}

# for a more complete view, see ?insert_labels
}
