% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{insert_labels}
\alias{insert_labels}
\title{Insert labels in factor variables (in portuguese).}
\usage{
insert_labels(
  file_path_origin,
  file_path_destiny,
  data_level = c("Escola", "Docente", "Gestor", "Turma", "Matricula"),
  add_variables = TRUE
)
}
\arguments{
\item{file_path_origin}{character, path to .rda file generated by function import_csv2rda()}

\item{file_path_destiny}{character, path to .rda file to be created, with labels in factors}

\item{data_level}{character, defines census data level according to data file imported ('Escola','Docente','Gestor','Turma','Matricula')}

\item{add_variables}{logical, if TRUE add some useful variables to the data.frame}
}
\value{
data.frame saved in .rda file
}
\description{
The dictionary used to label is from 2019 data
}
\details{
Works fine with 2019 data, and at least partially with 2015 data. Other years not tested yet.
}
\examples{
# you must firt download the .zip file. See ?download_microdata
# then you must unzip it and choose a data file (eg. ESCOLA.CSV)
# then you can run the code below, changing the first 3 lines as you wish
# note that 'data_level' must also be defined, in function insert_labels()

\dontrun{
file.original='~/YOUR_PATH/downloaded_data/ESCOLA.CSV'
file.imported='~/YOUR_PATH/temp_data.rda'
file.labelled='~/YOUR_PATH/censusData_ESCOLA.rda'

import_csv2rda(file_path_origin=file.original,
              file_path_destiny=file.imported)

insert_labels (file_path_origin=file.imported,
              file_path_destiny=file.labelled,
              data_level='Escola',
              add_variables=TRUE)}
}
